Refactor to use the unified task. Fixes #58
